% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloud9_operations.R
\name{delete_environment}
\alias{delete_environment}
\title{Deletes an AWS Cloud9 development environment}
\usage{
delete_environment(environmentId)
}
\arguments{
\item{environmentId}{[required] The ID of the environment to delete.}
}
\description{
Deletes an AWS Cloud9 development environment. If an Amazon EC2 instance is connected to the environment, also terminates the instance.
}
\section{Accepted Parameters}{
\preformatted{delete_environment(
  environmentId = "string"
)
}
}

\examples{
# 
\donttest{delete_environment(
  environmentId = "8d9967e2f0624182b74e7690ad69ebEX"
)}

}
