% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloud9_operations.R
\name{describe_environment_status}
\alias{describe_environment_status}
\title{Gets status information for an AWS Cloud9 development environment}
\usage{
describe_environment_status(environmentId)
}
\arguments{
\item{environmentId}{The ID of the environment to get status information about.}
}
\description{
Gets status information for an AWS Cloud9 development environment.
}
\examples{
# 
describe_environment_status(
  environmentId = "8d9967e2f0624182b74e7690ad69ebEX"
)

}
