% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloud9_operations.R
\name{list_environments}
\alias{list_environments}
\title{Gets a list of AWS Cloud9 development environment identifiers}
\usage{
list_environments(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{During a previous call, if there are more than 25 items in the list, only the first 25 items are returned, along with a unique string called a \emph{next token}. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.}

\item{maxResults}{The maximum number of environments to get identifiers for.}
}
\description{
Gets a list of AWS Cloud9 development environment identifiers.
}
\examples{
# 
list_environments()

}
