% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloud9_operations.R
\name{update_environment_membership}
\alias{update_environment_membership}
\title{Changes the settings of an existing environment member for an AWS Cloud9 development environment}
\usage{
update_environment_membership(
  environmentId = "string",
  userArn = "string",
  permissions = "read-write"|"read-only"
)
}
\arguments{
\item{environmentId}{[required] The ID of the environment for the environment member whose settings you want to change.}

\item{userArn}{[required] The Amazon Resource Name (ARN) of the environment member whose settings you want to change.}

\item{permissions}{[required] The replacement type of environment member permissions you want to associate with this environment member. Available values include:
\itemize{
\item \code{read-only}: Has read-only access to the environment.
\item \code{read-write}: Has read-write access to the environment.
}}
}
\description{
Changes the settings of an existing environment member for an AWS Cloud9 development environment.
}
\examples{
# 
update_environment_membership(
  environmentId = "8d9967e2f0624182b74e7690ad69ebEX",
  permissions = "read-only",
  userArn = "arn:aws:iam::123456789012:user/AnotherDemoUser"
)

}
