% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cloud9_operations.R
\name{describe_environments}
\alias{describe_environments}
\title{Gets information about AWS Cloud9 development environments}
\usage{
describe_environments(environmentIds)
}
\arguments{
\item{environmentIds}{[required] The IDs of individual environments to get information about.}
}
\description{
Gets information about AWS Cloud9 development environments.
}
\section{Accepted Parameters}{
\preformatted{describe_environments(
  environmentIds = list(
    "string"
  )
)
}
}

\examples{
# 
\donttest{describe_environments(
  environmentIds = list(
    "8d9967e2f0624182b74e7690ad69ebEX",
    "349c86d4579e4e7298d500ff57a6b2EX"
  )
)}

}
